realtime-voting/requirements.txt
streamlit==1.20.0
streamlit-autorefresh==0.1.4
pandas==2.1.1
matplotlib==3.8.0
kafka-python==2.0.2
pyspark==3.3.0
six==1.16.0
